pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tifaw*"]

[project]
name = "tifaw"
version = "0.1.0"
description = "Local AI desktop assistant & smart file organizer powered by Gemma 4"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "watchdog>=6.0",
    "pymupdf>=1.25",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "pillow>=11.0",
    "python-multipart>=0.0.18",
    "aiosqlite>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
tifaw = "tifaw.main:run"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
